Feat: 팀페이지 구현 - 레이아웃 완료
- style 코드 분리
- Detail page, Update page layout 구현
- banner 구현 (16:9)
