+++ .gitignore
@@ -7,6 +7,7 @@
 
 # testing
 /coverage
+/public/private
 
 # production
 /build
